backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
python-multipart
posthog
openai
pymupdf
python-docx
python-dotenv
email-validator
